New test SearchSubword focusing on the stackoverflow question
How to search some specific string or a word and there coordinates from
a pdf document in java
http://stackoverflow.com/questions/35937774/how-to-search-some-specific-string-or-a-word-and-there-coordinates-from-a-pdf-do
Variables.pdf
generated from Variables.docx

This test demonstrates how one can search for text parts with positions
in a PDF.

This method actually is quite crude as it expects the whole search term
to be forwarded to PDFTextStripper.writeString(String,
List<TextPosition>) in the same call. This is why the methods here are
called printSubwords(PDDocument, String) and findSubwords(PDDocument,
int, String) as internally PDFBox calls the portions forwarded here
"words", cf. the calling method
PDFTextStripper#writeLine(List<WordWithTextPositions>).

This class uses the helper class TextPositionSequence which implements
CharSequence for a list of TextPosition instances to allow for use as a
String-like object in case of character-oriented operations.