New test FillInForm focusing on the stackoverflow question
Apache PDFBox Form Fill TrueType text spacing issue
http://stackoverflow.com/questions/36926060/apache-pdfbox-form-fill-truetype-text-spacing-issue
FillFormField.pdf
https://www.dropbox.com/sh/b7ft1k0wfesob8s/AABnrTOgX26JlWCxl85jXns0a/FillFormField.pdf?dl=0

Indeed, the issue can be reproduced, it is due to a combination of two
factors:

A quirk of PDFBox when writing text - When writing text into a content
stream, PDFBox translates each Unicode codepoint into a name and looks
up that name in a map generating from the inverted font encoding. For
some encodings, though, there are two codes mapping to the name space,
and the inverted map maps back to only one of them, in the case at hand
the non-breaking variant. As both are expected to be typographically
identical, this should not be a problem. But:

Non-conformant font in the PDF - The font Impact in the PDF is defined
with width 176 for the normal space glyph and 750 for the nonbreaking
space glyph. Thus, they typographically differ vehemently.