New test CopyForm focusing on the stackoverflow question
PDFBox 2.0.2 How to copy PDTextFields between documents
http://stackoverflow.com/questions/39260500/pdfbox-2-0-2-how-to-copy-pdtextfields-between-documents

The NullPointerException can be reproduced. The problematic code line
eventually causing it is newField.getWidgets().add(newWidget): the
method getWidgets() merely returns a list with all the current widgets
of the field, it is not the list in which the field object itself keeps
its widgets. Thus, adding to this list does not attach the widget to the
field. Instead one needs to collect all the widgets in a new List and
eventually set this list as the list of field widgets using the
setWidgets method. But there are more issues in this code to resolve...

The nomenclature for methods here is indeed misleading: You can add a
page to a document using document.getPages().add(page) but you cannot
add widgets to a field using field.getWidgets().add(widget). IMO all
getXXXs across PDFBox should function similarly, either all allow using
add to add to the underlying document structure or none do.