New ExtractImages test testExtractPageImageResources10948 focusing on the stackoverflow question
How can I check if PDF page is image(scanned) by PDFBOX, XPDF
http://stackoverflow.com/questions/40531871/how-can-i-check-if-pdf-page-is-imagescanned-by-pdfbox-xpdf
10948.pdf
https://drive.google.com/file/d/0B9izTHWJQ7xlT2ZoQkJfbGRYcFE

The only special thing about the two images returned for the sample PDF is that one image is merely a mask used for the other image, and the other image is the actual image used on the PDF page. If one only wants the images immediately used in the page content, one also has to scan the page content.