New RotatePageContent tests focusing on the stackoverflow question
Rotate PDF around its center using PDFBox in java
http://stackoverflow.com/questions/40611736/rotate-pdf-around-its-center-using-pdfbox-in-java

Indeed, using testRotateLikeSagar (the code of the OP) the image mostly is rotated out of the page area.

testRotateCenter shows how to rotate the page content around the center of its crop box.

testRotateCenterScale shows how to rotate the page content around the center of its crop box and then crop it to make all previously visible content fit.

testRotateMoveBox shows how to rotate the page content and then move its crop box and media box accordingly to make it appear as if the content was rotated around the center of the crop box.

testRotateExpandBox shows how to rotate the page content and then set the crop box and media box to the bounding rectangle of the rotated page area.