New PlayWithHelloSign tools and tests focusing on the stackoverflow question
PDFBox: Remove a single field from PDF
http://stackoverflow.com/questions/41071142/pdfbox-remove-a-single-field-from-pdf
input.pdf
https://www.dropbox.com/s/oyv1vjyhkmao1t1/input.pdf?dl=0

The class HelloSignAnalyzer can be used to analyze a document filled by HelloSign (at least it can be used to analyze the example document and similar ones). 

The class HelloSignManipulator can be used to manipulate a document filled by HelloSign (at least it can be used to manipulate the example document and similar ones). It makes use of information dug up by the HelloSignAnalyzer class.

testAnalyzeInput applies the HelloSignAnalyzer to the sample document and outputs the found fields.

testClearAddress1Input applies the HelloSignManipulator to the sample document and clears the field var1001 (address1).

testClearZipStateDate2Input applies the HelloSignManipulator to the sample document and clears the fields var1004 (zip), var1003 (state), and date2.

testClearZipStateDate2SuccessivelyInput applies the HelloSignManipulator to the sample document multiple times and clears the fields var1004 (zip), var1003 (state), and date2, each field in a separate call.