New ValidateSignature tests focusing on the stackoverflow question
PDF Signature Validation
http://stackoverflow.com/questions/41116833/pdf-signature-validation
SignatureVlidationTest.pdf
https://drive.google.com/file/d/0BzEmZ9pRWLhPOUJSYUdlRjg2eEU/view?usp=sharing
pkcs7DetachedFailure.pdf
https://drive.google.com/file/d/0BzEmZ9pRWLhPblJJWXNuY2tneHM/view?usp=sharing

testValidateSignatureVlidationTestAdbePkcs7Sha1 (like the already checked-in testValidateSignatureVlidationTest) also ignores the SubFilter of the signature, in contrast to that other test, though, it is appropriate for signatures with SubFilter value adbe.pkcs7.sha1 which the example document has been signed with.

testValidateSignatureVlidationTestImproved and testValidatePkcs7DetachedFailureImproved use the method validateSignaturesImproved which is SubFilter-aware.

The reason for testValidatePkcs7DetachedFailureImproved failing here while the signature validation succeeds in Adobe Reader is that the signed attributes here are not properly DER encoded - the order in the set is wrong. Adobe Reader validates using the original, non-DER signed attributes representation while BouncyCastle here validates using a DER signed attributes representation.