New ExtractTextSections test testWang05a focusing on the stackoverflow question
How to read PDF departments(header,abstract,refrences) With PDFBox?
http://stackoverflow.com/questions/41654518/how-to-read-pdf-departmentsheader-abstract-refrences-with-pdfbox
Wang05a.pdf
http://merlot.usc.edu/csac-f06/papers/Wang05a.pdf

This test demonstrates a simple framework for extraction of semantic text sections which are recognizable by their characteristics of each line alone.

PDFTextSectionStripper is a PDFTextStripper subclass which is the hub of this framework.
TextSectionDefinition defines a section for extraction of semantic text sections in this framework.
TextSection represents a result section in the context of this framework.