New FillInFormSaveIncremental test testFillInSkierowanie3 focusing on the stack overflow question
PDFBox Form fill - saveIncremental does not work
http://stackoverflow.com/questions/42802996/pdfbox-form-fill-saveincremental-does-not-work
Skierowanie3.pdf
https://www.dropbox.com/s/mmjel976l6sjn4w/Skierowanie3.pdf?dl=0

The OP's code had to be changed in a number of ways. First of all all changed objects including objects building a chain to each of them starting at the catalog have to have NeedToBeUpdated set to true. This causes the usage rights signature also to be copied into the update section, so this signature has to be removed. Furthermore the form is a hybrid AcroForm/XFA form. As we only change the AcroForm entries, the xfa form also has to be removed.