New RemoveStrikeoutComment test focusing on the stack overflow question
PDFBox delete comment maintain strikethrough
https://stackoverflow.com/questions/45812696/pdfbox-delete-comment-maintain-strikethrough
only_fields.pdf
https://expirebox.com/files/3d955e6df4ca5874c38dbf92fc43b5af.pdf
https://file.io/DTvqhC

testRemoveLikeStephan, the OP's original code: Due to a bug in the
COSArrayList usage for page annotations, the indirect reference to the
annotation in question is not removed from the actual page annotations
array.

testRemoveLikeStephanImproved: The OP only wanted the comment removed, not the strike-through. Thus, we must not remove the annotation but merely the comment building attributes. So we do here.