New RectanglesOverText test testCoverTextByRectanglesMwbI201711 focusing on the stack overflow question
Text coordinates when stripping from PDFBox
https://stackoverflow.com/questions/46080131/text-coordinates-when-stripping-from-pdfbox
mwb_I_201711.pdf
https://download-a.akamaihd.net/files/media_mwb/b7/mwb_I_201711.pdf

This test applies the OP's code to his example PDF file and indeed,
there is an offset! This is due to the LegacyPDFStreamEngine method
showGlyph which manipulates the text rendering matrix to make the lower
left corner of the crop box the origin. In the current version of this
test, that offset is corrected, see below.