New ExtractVisibleText test focusing on the stack overflow question
PDFBox - Removing invisible text (by clip/filling paths issue)
https://stackoverflow.com/questions/47908124/pdfbox-removing-invisible-text-by-clip-filling-paths-issue
test.pdf (as testDmitryK.pdf)
https://drive.google.com/open?id=1xcZOusx3cEdZX4AT8QAVDqZe33YWla0H

Indeed, using the original PDFVisibleTextStripper implementation a lot of visible characters where dropped. This was due to the incorrect calculation of the `end` of the character baseline in the PDFVisibleTextStripper methods processTextPosition and deleteCharsInPath.

After patching those PDFVisibleTextStripper methods to make use of `end` only optionally, running the test with that option results in a decent extraction of visible text.