New CompareText test focusing on the stack overflow question
Comparing two PDF files text using PDFBox is failing eventhough both files are having same text
https://stackoverflow.com/questions/49256943/comparing-two-pdf-files-text-using-pdfbox-is-failing-eventhough-both-files-are-h
Table_Basic.pdf
https://drive.google.com/file/d/1SeqdWhZdp-wuc0sfR3tvaSIRlRBSYOxL/view?usp=sharing
Table_Basicf0313153905592.pdf
https://drive.google.com/file/d/1mFD2cPSSzY3OHTP07d2ZUosaI0ZyCpf2/view?usp=sharing

Indeed, there is a difference in output. It is due to different ToUnicode mappings of certain characters in a certain font. As the Unicode code points in question are in a private use area, this wasn't that apparent.