New AddContentToOCG test focusing on the stack overflow question
PDFbox - how can i add content to a layer?
https://stackoverflow.com/questions/49498074/pdfbox-how-can-i-add-content-to-a-layer

This test show how add content to an optional content group which may or may not already exist.

This is the PDFBox 2 equivalent of the PDFBox 1 test with the same name in response to

How do I make modifications to existing layer(Optional Content Group) in pdf?
http://stackoverflow.com/questions/43275212/how-do-i-make-modifications-to-existing-layeroptional-content-group-in-pdf