New CreateSignature test focusing on the stack overflow question
PDFBox 2.0.8 issue while signing document
https://stackoverflow.com/questions/50224181/pdfbox-2-0-8-issue-while-signing-document

This test checks the OP's code which essentially is a copy of the PDFBox example class CreateVisibleSignature with an extra method.

The cause was that the OP chose a PDDocument.addSignature overload for external signing while implementing otherwise internal signing. Choosing an appropriate overload fixed the problem.