New ExtractImageLocations tests focusing on the stack overflow question
Why does PDFBox return image dimension of size 0 x 0
https://stackoverflow.com/questions/50576614/why-does-pdfbox-return-image-dimension-of-size-0-x-0
TopSecret.pdf
https://files.fm/u/udvgszyv#_

testExtractLikeHelloWorldFromTopSecret effectively is the code used by the OP. Indeed, the image is not found by the PrintImageLocations class. This is due to the image being drawn in a tiling pattern which then is used to fill an area in the page content and PrintImageLocations not descending into patterns.

testExtractLikeHelloWorldImprovedFromTopSecret uses an improved version of the PrintImageLocations class: PrintImageLocationsImproved. The improvement makes the class descent into a pattern definition whenever a filling operation is encountered with a pattern color.
This is not perfect as it does neither properly restrict to the area actually filled nor return multiple finds for an area large enough to require multiple pattern tiles to fill. Nonetheless, it returns an image match.