New ExtractBoxedText test focusing on the stack overflow question
Extracting text from pdf (java using pdfbox library) from a table's rows with different heights
https://stackoverflow.com/questions/51380677/extracting-text-from-pdf-java-using-pdfbox-library-from-a-tables-rows-with-di
Programare-licenta-5-Iulie-2018_1.pdf
https://www.info.uvt.ro/wp-content/uploads/2018/07/Programare-licenta-5-Iulie-2018_1.pdf

PdfBoxFinder is a stream engine class which determines the lines framing table cells. It is implemented to recognize lines created like in the example PDF shared by the OP, i.e. lines drawn as long thin filled rectangles. It is easily possible to generalize this for frame lines drawn differently, c.f. the method processPath().

testExtractBoxes is a first check of the PdfBoxFinder. It merely outputs the locations of identified horizontal and vertical lines.

testDrawBoxes draws a grid on the test file representing the boxes found by the PdfBoxFinder.

testExtractBoxedTexts feeds the regions found by the PdfBoxFinder into a PDFTextStripperByArea and extracts the text of the areas in questions.