New DetermineBoundingBox tests focusing on the stack overflow question
How do determine location of actual PDF content with PDFBox?
https://stackoverflow.com/questions/52821421/how-do-determine-location-of-actual-pdf-content-with-pdfbox

This class tests the BoundingBoxFinder by applying it to a number of
miscellaneous PDFs and stroking the determined boxes.

The BoundingBoxFinder determines the bounding box of the static content
of a page. Beware, it is not very sophisticated; in particular it does
not ignore invisible content like a white background rectangle, text
drawn in rendering mode "invisible", arbitrary content covered by a
white filled path, white parts of bitmap images, ... Furthermore, it
ignores clip paths.