New OptimizeAfterMerge test focusing on the stack overflow question
Ho to reduce the size of merged PDF A1/b Files with pdfbox or other java
library
https://stackoverflow.com/questions/53420344/ho-to-reduce-the-size-of-merged-pdf-a1-b-files-with-pdfbox-or-other-java-library
dummy.pdf
https://datentransfer.sparkassenverlag.de/my/transfers/5q8eskgne52npemx8kid7728zk1hq3f993dfat8his

testOptimizeDummy applies the method optimize(PDDocument) to the
document supplied by the OP to demonstrate how to remove identical
duplicates (in particular streams) using PDFBox.

optimize with its helper methods and classes attempts to identify
identical objects in a PDF and remove all but a single instance of them.
This is meant to target in particular streams but other complex objects
(arrays and dictionaries) are also affected.
This method might be a bit over-eager in its removal. E.g. in case of
separate pages Adobe Reader expects different objects for different
pages. This method, though, might collaps identically built pages to a
single objects. Thus, this method probably needs to be tamed a bit.