New OptimizeAfterMerge test focusing on the stack overflow question
Ho to reduce the size of merged PDF A1/b Files with pdfbox or other java
library
https://stackoverflow.com/questions/53420344/ho-to-reduce-the-size-of-merged-pdf-a1-b-files-with-pdfbox-or-other-java-library
dummy.pdf
https://datentransfer.sparkassenverlag.de/my/transfers/5q8eskgne52npemx8kid7728zk1hq3f993dfat8his

testOptimizeLikeSchowaveDummy is the code the OP himself posted as his
solution.

This only works if (a) all font programs embedded for the same name
indeed are identical, and if (b) all fonts to consider are in the
immediate page resources, not the resources of some referred to xobject
or pattern. If those conditions are fulfilled, though, it very likely is
much faster than the approach in optimize(PDDocument). For the example
file provided by the OP, its result is nearly as small.