New FillImageField test focusing on the stack overflow question
Setting image form field
https://stackoverflow.com/questions/54081053/setting-image-form-field
test_wo_img.pdf
https://drive.google.com/file/d/1balEQf_P6_KPsJGMGUH3jejLVL5wHFds/view?usp=sharing

This test uses Renat Gatin's AcroFormPopulator to solve the task of the
OP.

Setting the populateAndCopy parameter flatten to true also reveals a bug
in form flattening of PDFBox.

The AcroFormPopulator utility class is from Renat Gatin's answer to the
stack overflow question

How to insert image programmatically in to AcroForm field using java
PDFBox?
https://stackoverflow.com/questions/46799087/how-to-insert-image-programmatically-in-to-acroform-field-using-java-pdfbox

A flag parameter flatten was added to populateAndCopy to allow deciding
whether or not to flatten the form.