New RepeatingTableHeaders tests focusing on the stack overflow question
PDFBOX - header in all pages using easytable
https://stackoverflow.com/questions/54233886/pdfbox-header-in-all-pages-using-easytable

createTwoPageTable is the code from the test routine the OP referred to.
Indeed, it does not create the header rows on each page, there simply is
no explicit support for table headers in easytable. But it is possible
to add repeating headers nonetheless...

createTwoPageTableRepeatingHeader is derived from the code from the test
routine the OP referred to. It has been changed to put the table headers
in a separate table which is drawn again and again. Indeed, it does
create the header rows on each page, unlike the original code.