New DetermineTextAndBackgroundColors test focusing on the SO question
Background color is incorrect on first page for some reason
https://stackoverflow.com/questions/54637141/background-color-is-incorrect-on-first-page-for-some-reason
test.pdf as "test3DmitryK.pdf"
https://drive.google.com/open?id=1LXOuGmk67hQRwLXi5GBx-5n3-CJ_NV7h

The underlying issue is a bug in
org.apache.pdfbox.pdmodel.common.function.PDFunctionType0#eval(float[])
- it changes values in its parameter array which is something other
PDFBox does not expect.

PdfToTextInfoConverter is the OP's class skeleton for a text extractor
that recognizes text drawn in the same color as its background. The
problem is, though, that the bug mentioned above changes this class'
underlying data. But there also is an unrelated bug here which doesn't
play a role in the question at hand, though.