New SetRichText test focusing on the stack overflow question
acroform field.setRichTextValue is not working
https://stackoverflow.com/questions/54988511/acroform-field-setrichtextvalue-is-not-working
form.pdf as "formBee.pdf"
https://drive.google.com/open?id=1jFbsYGFOnx8EMiHgDsE8LQtfwJHSa5Gh

After correction of a few details, the code kinda runs. In particular it
is necessary to use PDF style rich text (and not LaTeX richtext package
instructions), to set the flag NeedAppearances to <code>true</code>, and
to provide a V value equal to the RV value without markup.

It only "kinda" runs because the OP also wants to flatten the form which
doesn't work as PDFBox does not create appearance streams based on the
rich text value.