New ExtractLinesWithDir tests focusing on the stack overflow question
PDFBox - Line / Rectangle extraction
https://stackoverflow.com/questions/55166990/pdfbox-line-rectangle-extraction
LineRotationTest.pdf
https://github.com/yashodhan19/PDF/blob/master/LineRotationTest.pdf

testExtractLineRotationTestLikeYashodhanJoglekar essentially is the LineCatcher main method code posted by the OP, merely dressed up as a unit test. The results are coordinates of the bottom left point of the line bounding boxes without consideration of page rotation.

testExtractLineRotationTestWithDir in contrast to the OP's original code attempts to extract the coordinates as the OP wants it, i.e. the coordinates of the top left point of the line bounding boxes on the rotated page, the origin in the upper left corner of the page.